The Lognormals Module

Missing docstring.

Missing docstring for Lognormals. Check Documenter's build log for details.

Module Index

Lognormals.AbstractMoments

Lognormals.plusTwo

Detailed API

Lognormals.AbstractMoments — Type

AbstractMoments

A representation of statistical moments of a distribution

The following functions are supported

length(M::AbstractMoments): get the number of defined moments

The following getters return a single moment or throw an error if the moment has not been defined

mean(M::AbstractMoments): get the mean

var(M::AbstractMoments): get the variance

skewness(M::AbstractMoments): get the variance

kurtosis(M::AbstractMoments): get the variance

getindex(M::AbstractMoments,i): get the ith moment, i.e. indexin [i]

Moments-objects be converted to AbstractArray by

convert(AbstractArray, M::AbstractMoments)

The basic implementation 'Moments' is immutable and convert(AbstractArray, M::Moments) returns an SArray{N,T}

Examples

M = Moments(1,0.2)
length(M) == 2
var(M) == M[2]
ismissing(kurtosis(M))
typeof(convert(AbstractArray,M)) <: AbstractArray

Lognormals.plusTwo — Method

plusTwo(x)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5

StatsBase.fit — Method

fit(Distribution, M::Moments)

Fit a statistical distribution to its moments

Arguments

Distribution: The type of distribution to fit

M: The moments of the distribution

Examples

D = fit(LogNormal, Moments(3.2,4.6))