Standard error of the mean of correlated series

Effective number of observations

For uncorrelated series, the variance of the mean decreases by the number of records compared to the variance of a single record.

\[Var(\bar{x}) = {Var(x) \over n}\]

For correlated series, the effective number of observations is defined as to give the same scaling.

\[Var(\bar{x}) = {Var(x) \over n_{eff}}\]

Since $n_{eff} < n$ for positive correlations, the uncertainty of the mean, i.e. standard deviation of the mean is larger than that for an uncorrelated series.

LogNormals.effective_n_cor — Method

effective_n_cor(x, acf::AbstractVector; exactmissing::Bool=true)

Compute the number of effective observations for an autocorrelated series.

The formula in Zieba has been extended for missing values:

\[n_{eff} = \frac{n_F}{1+{2 \over n_F} \sum_{k=1}^{min(n-1,n_k)} (n-k-m_k) \rho_k}\]

where $n$ is the number of total records, $n_F$ is the number of finite records, $n_k$ is the nummber of components in the used autocorrelation function ($n-1$ if not estimated from the data) ,$\rho_k$ is the correlation, and $m_k$ is the number of pairs that contain a missing value for lag $k$.

Arguments

x: An iterator of a series of observations

acf: AutocorrelationFunction starting from lag 0

Optional Arguments

exactmissing: set to false to speed up computation (omitting count_forlags missing pairs) but get a positively biased result with increasing bias with the number of missings. This leads to a subsequent underestimated uncertainty of the sum or the mean.

Examples

acf0 = [1,0.4,0.1]
Sigma = cormatrix_for_acf(100, acf0);
# 100 random variables each Normal(1,1)
dmn = MvNormal(ones(100), Symmetric(Sigma));
x = allowmissing(rand(dmn));    
x[11:20] .= missing
neff = effective_n_cor(x, acf0)
neff < 90
neff_biased = effective_n_cor(x, acf0; exactmissing=false)
neff_biased > neff

Standard error of the mean of a correlated series

LogNormals.sem_cor — Method

sem_cor(x; exactmissing::Bool=true)
sem_cor(x, acf::AbstractVector; exactmissing::Bool=true)

Estimate the standard error of the mean of an autocorrelated series: $Var(\bar{x}) = {Var(x) \over n_{eff}}$.

Arguments

x: An iterator of a series of observations

acf: AutocorrelationFunction starting from lag 0. If not specified it is estimated from the data using provided value of exactmissing.

Optional Arguments

exactmissing: passed to effective_n_cor allowing to speed up computation, but leading to a negatively biased result with bias increasing with number of missings.

neff: may provide a precomputed number of observations for efficiency.

Variance of a correlated series

LogNormals.var_cor — Method

var_cor(x, acfe; exactmissing::Bool=true, neff=nothing)

Estimate the variance for an autocorrelated series.

Zieba 2011 provide the following formula:

\[Var(x) = \frac{n_{eff}}{n (n_{eff}-1)} \sum \left( x_i - \bar{x} \right)^2 = {(n-1) n_{eff} \over n (n_{eff}-1)} Var_{uncor}(x)\]

Arguments

x: An iterator of a series of observations

acf: AutocorrelationFunction starting from lag 0

Optional Arguments

exactmissing: passed to effective_n_cor allowing to speed up computation, leading to a slighly positively biased result with small neff and many missings.

neff: may provide a precomputed number of observations for efficiency.

Effective autocorrelation function

LogNormals.autocor_effective — Method

autocor_effective(x, acf = autocor(x))

Estimate the effective autocorrelation function for series x.

Arguments

x: An iterator of a series of observations

acf: AutocorrelationFunction starting from lag 0

Notes

The effect autocorrelation function are the first coefficients of the autocorrelation function up to before the first negative coefficient.

According to Zieba 2011 using this effective version rather the full version when estimating the autocorrelationfunction from the data yields better result for the standard error of the mean (sem_cor).

Optional argument acf allows the caller to provide a different estimate, of autocorrelation function e.g. using exactmissing=false in autocor.

Autocorrelation of a series with missing values

StatsBase.autocor — Method

autocor(x::AbstractVector{Union{Missing,<:T}, lags; dmean, exactmissing=true}

Estimate the autocorrelation function accounting for missing values.

Arguments

exactmissing: If set to true then devide the sum in the formula of the exepected value in the formula for the correlation at lag k by n - nmissing instead of n, where nimissing is the number of records where there is a missing either in the original vector or its lagged version (see count_forlags).

If exactmissing=false then the computation is faster, but it is more strongly biased low with increasing number of missings. Note that StatsBase.autocor uses devision by n instead of 'n-k', the true length of the vectors correlated at lag k resulting in low-biased correlations of higher lags for numerical stability reasons.

LogNormals.count_forlags — Method

count_forlags(pred, x, lags)
count_forlag(pred, x, k::Integer)

Count the number of records in where each pair of records of lag k fulfil a predicate.

Arguments

pred::Function(x_i,x_iplusk)::Bool: The predicate to be applied to each pair

x: The series whose lags are inspected.

lags: An iterator of Integer lag sizes

k: A single lag.