The LogNormals Module

Missing docstring.

Missing docstring for LogNormals. Check Documenter's build log for details.

Module Index

LogNormals.AbstractMoments

Detailed API

LogNormals.AbstractMoments — Type

AbstractMoments

A representation of statistical moments of a distribution

The following functions are supported

length(M::AbstractMoments): get the number of defined moments

The following getters return a single moment or throw an error if the moment has not been defined

mean(M::AbstractMoments): get the mean

var(M::AbstractMoments): get the variance

skewness(M::AbstractMoments): get the variance

kurtosis(M::AbstractMoments): get the variance

getindex(M::AbstractMoments,i): get the ith moment, i.e. indexing m[i]

The basic implementation 'Moments' is immutable and convert(AbstractArray, M::Moments) returns an SArray{N,T}

Examples

M = Moments(1,0.2)
length(M) == 2
var(M) == M[2]
ismissing(kurtosis(M))
typeof(convert(AbstractArray,M)) <: AbstractArray

StatsBase.fit — Method

fit(Distribution, m::Moments)

Fit a statistical distribution to its moments

Arguments

Distribution: The type of distribution to fit

m: The moments of the distribution

Notes

This can be used to approximate one distribution by another

Examples

D = fit(LogNormal, Moments(3.2,4.6))
D = fit(LogNormal, moments(Normal(3,1)))
plot(D); lines(!Normal(3,1))

StatsBase.fit — Method

fit(Distribution, lower::QuantilePoint, upper::QuantilePoint)

Fit a statistical distribution to a set of quantiles

Arguments

Distribution: The type of distribution to fit

lower: lower QuantilePoint (p,q)

upper: upper QuantilePoint (p,q)

Notes

Several macros help to construct QuantilePoints @qp(p,q) QuantilePoint(p, q) @qpll(q0025) quantile at very low p = 0.025 @qpl(q005) quantile at low p = 0.05 @qpm(median) quantile at median: p = 0.5 @qpu(q095) quantile at high p = 0.95 @qpuu(q0_975) quantile at very high p = 0.975

Examples

DN = fit(Normal, @qp_m(3), @qp_uu(5))
D = fit(LogNormal, @qp_m(3), @qp_uu(5))
quantile.(D, [0.5, 0.975]) ≈ [3,5]

StatsBase.fit — Method

fit(Distribution, val::Real, qp::QuantilePoint, ::Val{stats} = Val(:mean))

Fit a statistical distribution to a quantile and given statistics

Arguments

Distribution: The type of distribution to fit

val: The value of statistics

qp: QuantilePoint(p,q)

stats Which statistics to fit: defaults to Val(:mean). Alternatives are: Val(:mode), Val(:median)

Examples

D = fit(LogNormal, 5, @qp_uu(14))
mean(D) ≈ 5 && quantile(D, 0.975) 
D = fit(LogNormal, 5, @qp_uu(14), Val(:mode))
mode(D) ≈ 5 && quantile(D, 0.975)