Aggreatation of random variables

Sum and Mean of several correlated random variables

Base.sum — Method

sum(dv::AbstractDistributionVector; <keyword arguments>)
sum(dv::AbstractDistributionVector, corr; <keyword arguments>)
sum(dv::AbstractDistributionVector, acf; <keyword arguments>)

mean(dv::AbstractDistributionVector; <keyword arguments>)
mean(dv::AbstractDistributionVector, corr; <keyword arguments>)
mean(dv::AbstractDistributionVector, acf; <keyword arguments>)

Compute the distribution of the sum or mean of correlated random variables.

Arguments

dv: The vector of distributions, see AbstractDistributionVector

An optional second arguments supports correlation between random variables.

corr::Symmetric: correlation matrix, or

acf::AutoCorrelationFunction: coefficients of the AutoCorrelationFunction

Keyword arguments:

skipmissings: Set to Val(true) to conciously care for missings in dv.

isgapfilled::AbstractVector{Bool}: set to true for records that should contribute to the sum but not to the decrease of relative uncertainty with increasing number of records, e.g. for missing records that have been estimated (gapfilled).

The sums of correlated variables require extra allocation and support an additional keyword parameter

storage: a mutable AbstractVector{eltype(D)} of length of dv that provides storage space to avoid additional allocations.

Helpers

If correlations are only dependent on the distance of records, one can specify these correlation by a vector starting with distance, i.e. lag, 1.

LogNormals.AutoCorrelationFunction — Type

AutoCorrelationFunction{T}

A representation of the autocorrelation function.

It supports accessing the coeficients starting from lag 0 by

coef(acf::AutoCorrelationFunction): implements StatsBase.coef

coef(acf::AutoCorrelationFunction, lag::Integer): coefficient for lag

Wrapping the vector of coefficients into its own type helps avoiding method ambiguities.

Examples

using StatsBase: coef
acf = AutoCorrelationFunction([1,0.4,0.1])
coef(acf) == [1,0.4,0.1]
coef(acf,1) == 0.4